@@ -20,18 +20,25 @@ long_description_content_type = text/markdown
 [options]
 packages = motioneye
 include_package_data = False
-python_requires = >=3.7
+python_requires = >=3.9
 install_requires =
     tornado>=6.0,<7.0
     jinja2>=3.0,<4.0
-    pillow>=8.0,<11.0
+    pillow>=8.0,<12.0
     pycurl>=7.43,<8.0
     babel>=2.9,<3.0
     boto3>=1.20,<2.0
     opencv-python>=4.5,<5.0
-    face_recognition>=1.3,<2.0
     paho-mqtt>=1.5,<2.0
-    psutil>=5.8,<6.0
+    psutil>=5.8,<7.0
+
+[options.extras_require]
+# Face recognition is optional due to heavy dlib dependency
+face_recognition =
+    face_recognition>=1.3,<2.0
+# Full installation with all features
+all =
+    face_recognition>=1.3,<2.0
 
 [options.package_data]
 motioneye = extra/*, static/*.*, static/*/*, templates/*, scripts/*, controls/*, handlers/*, utils/*, locale/*/LC_MESSAGES/*.mo
