@@ -129,7 +129,7 @@ def _exec_monitor_command(command):
     try:
         interval = int(err)
 
-    except:
+    except (ValueError, TypeError):
         interval = DEFAULT_INTERVAL
 
     out = out.strip()
