@@ -46,7 +46,7 @@ def _get_os_version_lsb_release():
 
         return name, version
 
-    except:
+    except (OSError, ValueError, IndexError):
         return _get_os_version_uname()
 
 
@@ -59,7 +59,7 @@ def _get_os_version_uname():
 
         return name, version
 
-    except:
+    except (OSError, subprocess.CalledProcessError, IndexError):
         return 'Linux', ''  # most likely :)
 
 
@@ -71,7 +71,7 @@ def compare_versions(version1, version2):
         try:
             return int(n)
 
-        except:
+        except (ValueError, TypeError):
             return 0
 
     version1 = [int_or_0(n) for n in version1.split('.')]
